+++ TShockAPI/Bouncer.cs
@@ -38,6 +38,7 @@ namespace TShockAPI
 	{
 		internal Handlers.SendTileSquareHandler STSHandler { get; set; }
 		internal Handlers.NetModules.NetModulePacketHandler NetModuleHandler { get; set; }
+    internal Handlers.EmojiHandler EmojiHandler { get; set; }
     internal Handlers.LandGolfBallInCupHandler LandGolfBallInCupHandler { get; set; }
 
 		/// <summary>Constructor call initializes Bouncer and related functionality.</summary>
@@ -50,6 +51,9 @@ namespace TShockAPI
 			NetModuleHandler = new Handlers.NetModules.NetModulePacketHandler();
 			GetDataHandlers.ReadNetModule += NetModuleHandler.OnReceive;
 
+			EmojiHandler = new Handlers.EmojiHandler();
+			GetDataHandlers.Emoji += EmojiHandler.OnReceiveEmoji;
+      
 			LandGolfBallInCupHandler = new Handlers.LandGolfBallInCupHandler();
 			GetDataHandlers.LandGolfBallInCup += LandGolfBallInCupHandler.OnLandGolfBallInCup;